Next_ERA_Version/VPP_Core/dispatch_service/requirements.txt
# Core Dependencies
fastapi>=0.68.0
pydantic>=1.8.2
uvicorn>=0.15.0
python-dotenv>=0.19.0

# Messaging
kafka-python>=2.0.2
aiokafka>=0.7.2

# Database
sqlalchemy>=1.4.23
databases>=0.5.2
asyncpg>=0.24.0  # PostgreSQL driver

# Optimization
numpy==1.21.0
scipy==1.7.0
pulp>=2.7.0  # Linear programming solver
pandas==1.3.0  # Data manipulation
cvxpy>=1.3.0  # Convex optimization
scikit-learn>=1.2.0  # Machine learning for forecasting

# Monitoring
prometheus-client>=0.11.0
opentelemetry-api>=1.4.1
opentelemetry-sdk>=1.4.1
opentelemetry-instrumentation-fastapi>=0.24b0

# Testing
pytest>=6.2.5
pytest-asyncio>=0.15.1
httpx>=0.19.0  # Async HTTP client

# Utilities
python-json-logger>=2.0.2
tenacity>=8.0.1  # Retry logic
structlog>=21.1.0  # Structured logging
pytz>=2023.3  # Timezone handling 